refa: initializeStore에 직접createAppStore 로직 추가
